Refactor `FakeMarketRepository` implementation and update tests to use it in `MarketDetailScreenTest`
* Moved `FakeMarketRepository` from `TestModule.kt` to `MarketRepository.kt` for better organization and reusability.
* Updated `MarketDetailScreenTest` to utilize the new `FakeMarketRepository` instead of an inline implementation, enhancing test clarity and maintainability.
